added create arc button

+++ src/app/designer/designer.component.ts
@@ -1,9 +1,10 @@
 import { AfterViewInit, Component, ElementRef, NgZone, ViewChild } from '@angular/core';
 import * as PIXI from 'pixi.js';
-import { PlaceReference as PlaceReference } from './placeReference';
+import { PlaceReference as PlaceReference } from './entities/placeReference';
 import {v4 as uuidv4} from 'uuid';
 import { Global} from './../globals'
-import { ArcReference } from './arcReference';
+import { ArcReference } from './entities/arcReference';
+import {  ChangeDetectorRef } from '@angular/core';
 
 @Component({
     selector: 'app-designer',
@@ -11,26 +12,30 @@ import { ArcReference } from './arcReference';
     styleUrls: ['./designer.component.css']
 })
 export class DesignerComponent implements AfterViewInit  {
-    @ViewChild('divParent') div: ElementRef
-    canvas: ElementRef
-    placeReferenceList: PlaceReference[] = new Array()
+    public arcBtnIsVisible: boolean = false
+    public createArcInProgress: boolean = false
+    public arcSourceNode: PlaceReference
 
-    constructor(private ngZone: NgZone) {}
+    @ViewChild('pixiCanvasContainer') private div: ElementRef
+    private placeReferenceList: PlaceReference[] = new Array()
+
+    constructor(private ngZone: NgZone, private cdr: ChangeDetectorRef) {}
 
     ngAfterViewInit(): void {
         this.ngZone.runOutsideAngular(() => {
-        // init application
-        Global.app = new PIXI.Application({
-            backgroundColor: 0x0d6efd
-        });
+            // init application
+            Global.app = new PIXI.Application({
+                backgroundColor: 0x0d6efd
+            });
 
-        this.div.nativeElement.innerHTML = ''
-        this.div.nativeElement.appendChild(Global.app.view)
+            this.div.nativeElement.innerHTML = ''
+            this.div.nativeElement.appendChild(Global.app.view)
+            this.adjustCanvasSize()
 
-        // load XML file
-        let designerData = localStorage.getItem("designerData")
-        if(designerData) this.loadDataFromLocal(designerData)
-        else this.createNewXMLDocument()
+            // load XML file
+            let designerData = localStorage.getItem("designerData")
+            if(designerData) this.loadDataFromLocal(designerData)
+            else this.createNewXMLDocument()
         });
     }
 
@@ -56,11 +61,12 @@ export class DesignerComponent implements AfterViewInit  {
                     .getElementsByTagName("graphics")[0]
                     .getElementsByTagName("position")[0]
                     .getAttribute("y")),
-                place
+                String(place
                     .getElementsByTagName("name")[0]
                     .getElementsByTagName("text")[0]
-                    .textContent as string,
-                false
+                    .textContent),
+                false,
+                this
                 ))
         })
 
@@ -70,15 +76,14 @@ export class DesignerComponent implements AfterViewInit  {
             .getElementsByTagName("net")[0]
             .getElementsByTagName("page")[0]
             .getElementsByTagName("arc")).forEach(arc => {
-                let tmpArc;
-                let sourceRef = this.placeReferenceList.find(el => el.id == String(arc.getAttribute("source")))
-                if(sourceRef) {
-                    tmpArc = new ArcReference(String(arc.getAttribute("source")), String(arc.getAttribute("target")), sourceRef.sprite)
-                    sourceRef.arcReferenceList.push(tmpArc)
-                }
-
-                let targetRef = this.placeReferenceList.find(el => el.id == String(arc.getAttribute("target")))
-                if(targetRef && tmpArc) targetRef.arcReferenceList.push(tmpArc)
+                this.addArc(String(arc.getAttribute("id")), 
+                String(arc.getAttribute("source")), 
+                String(arc.getAttribute("target")), 
+                String(arc
+                .getElementsByTagName("inscription")[0] 
+                .getElementsByTagName("text")[0] 
+                .textContent),
+                false)
             })
     }
 
@@ -92,13 +97,55 @@ export class DesignerComponent implements AfterViewInit  {
     }
   
     addObject(){
-        new PlaceReference(uuidv4(), false, 30, 30, "Test", true)
+        this.placeReferenceList.push(new PlaceReference(uuidv4(), false, 30, 30, "Test", true, this))
+    }
+    
+    addArc(id: string, sourceId: string, targetId: string, textValue: string, saveInXml: boolean){
+        let tmpArc;
+        let sourceRef = this.placeReferenceList.find(el => el.id == sourceId)
+        if(sourceRef) {
+            tmpArc = new ArcReference(id, sourceId, targetId, textValue, sourceRef.sprite, saveInXml)
+            sourceRef.arcReferenceList.push(tmpArc)
+        }
+
+        let targetRef = this.placeReferenceList.find(el => el.id == targetId)
+        if(targetRef && tmpArc) targetRef.arcReferenceList.push(tmpArc)
+
+        this.createArcInProgress = false
+        this.deactivateCreateArcBtn()
     }
 
     deleteLocalStorage(){
-        localStorage.clear()
-        if(Global.app) Global.app.destroy()
-        this.ngAfterViewInit()
+        if (confirm('Are you sure to delete all nodes and start from scratch?')) {
+            localStorage.clear()
+            if(Global.app) Global.app.destroy()
+            this.ngAfterViewInit()
+        }
+    }
+
+    public activateCreateArcBtn(sourceNode: PlaceReference){
+        if(this.arcSourceNode) this.arcSourceNode.sprite.tint = 0xFFFFFF // reset old node tint
+        
+        this.arcSourceNode = sourceNode
+        this.arcSourceNode.sprite.tint = 0x71beeb
+        this.arcBtnIsVisible = true
+        this.cdr.detectChanges();
+    }
+
+    public deactivateCreateArcBtn(){
+        if(this.arcSourceNode) this.arcSourceNode.sprite.tint = 0xFFFFFF // reset node tint
+
+        this.arcBtnIsVisible = false
+        this.cdr.detectChanges();
+    }
+
+    startCreateArc(){
+        this.createArcInProgress = true
+        this.cdr.detectChanges();
+    }
+
+    adjustCanvasSize(){
+        Global.app.renderer.resize(this.div.nativeElement.offsetWidth, this.div.nativeElement.offsetHeight)
     }
 
     ngOnDestroy(): void {
